[MOV] data_cleanup: move the code to the correct folder
Task: 6384868
