+++ src/components/side_panel/remove_duplicates/remove_duplicates.ts
@@ -2,8 +2,8 @@ import { useProps } from "@odoo/owl";
 import { numberToLetters } from "../../../helpers/coordinates";
 import { zoneToDimension } from "../../../helpers/zones";
 import { Component } from "../../../owl3_compatibility_layer";
-import { DataCleanupStore } from "../../../plugins/ui_feature/data_cleanup";
 import { useLocalStore } from "../../../store_engine/store_hooks";
+import { DataCleanupStore } from "../../../stores/data_cleanup_store";
 import { _t } from "../../../translation";
 import { SpreadsheetChildEnv } from "../../../types/spreadsheet_env";
 import { Store } from "../../../types/store_engine";
